UAS-101 endpoint for get the account plan done

+++ src/main/kotlin/ucb/accounting/backend/dao/repository/AccountGroupRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ucb.accounting.backend.dao.AccountGroup
 
 interface AccountGroupRepository: JpaRepository<AccountGroup, Long> {
-    fun findAllByCompanyIdAndAccountCategoryIdAndStatusIsTrue (companyId: Long, accountCategoryId: Long): List<AccountGroup>
+    fun findAllByCompanyIdAndAccountCategoryIdAndStatusIsTrue (companyId: Int, accountCategoryId: Int): List<AccountGroup>
 }
