+++ src/main/kotlin/ucb/accounting/backend/dao/repository/AccountRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ucb.accounting.backend.dao.Account
 
 interface AccountRepository: JpaRepository<Account, Long> {
-    fun findAllByCompanyIdAndAccountSubgroupIdAndStatusIsTrue (companyId: Long, accountSubgroupId: Long): List<Account>
+    fun findAllByCompanyIdAndAccountSubgroupIdAndStatusIsTrue (companyId: Int, accountSubgroupId: Int): List<Account>
 }
