+++ src/main/kotlin/ucb/accounting/backend/dao/repository/AccountSubGroupRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ucb.accounting.backend.dao.AccountSubgroup
 
 interface AccountSubGroupRepository: JpaRepository<AccountSubgroup, Long> {
-    fun findAllByCompanyIdAndAccountGroupIdAndStatusIsTrue (companyId: Long, accountGroupId: Long): List<AccountSubgroup>
+    fun findAllByCompanyIdAndAccountGroupIdAndStatusIsTrue (companyId: Int, accountGroupId: Int): List<AccountSubgroup>
 }
