+++ src/main/kotlin/ucb/accounting/backend/dao/repository/SubAccountRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ucb.accounting.backend.dao.Subaccount
 
 interface SubAccountRepository: JpaRepository<Subaccount, Long> {
-    fun findAllByCompanyIdAndAccountIdAndStatusIsTrue (companyId: Long, accountId: Long): List<Subaccount>
+    fun findAllByCompanyIdAndAccountIdAndStatusIsTrue (companyId: Int, accountId: Int): List<Subaccount>
 }
